+++ bin/_test-helpers.sh
@@ -439,7 +439,7 @@ run_cni-calico-deep_test() {
 run_helm-deep_test() {
   local tests=()
   setup_helm
-  helm_multicluster_chart="$( cd "$bindir"/.. && pwd )"/charts/linkerd2-multicluster
+  helm_multicluster_chart="$( cd "$bindir"/.. && pwd )"/multicluster/charts/linkerd2-multicluster
   run_test "$test_directory/install_test.go" --helm-path="$helm_path" --helm-chart="$helm_chart" \
   --helm-release="$helm_release_name" --multicluster-helm-chart="$helm_multicluster_chart" \
   --multicluster-helm-release="$helm_multicluster_release_name"
