+++ cli/Dockerfile-bin
@@ -15,6 +15,7 @@ WORKDIR /linkerd-build
 COPY cli cli
 COPY charts charts
 COPY jaeger jaeger
+COPY multicluster multicluster
 
 COPY controller/k8s controller/k8s
 COPY controller/api controller/api
@@ -26,6 +27,7 @@ RUN mkdir -p /out
 # TODO: `go generate` does not honor -mod=readonly
 RUN go generate -mod=readonly ./pkg/charts/static
 RUN go generate -mod=readonly ./jaeger/static
+RUN go generate -mod=readonly ./multicluster/static
 
 # Cache builds without version info
 RUN CGO_ENABLED=0 GOOS=darwin go build -o /out/linkerd-darwin -tags prod -mod=readonly -ldflags "-s -w" ./cli
