+++ cli/cmd/uninstall.go
@@ -3,11 +3,10 @@ package cmd
 import (
 	"context"
 	"fmt"
-	"io"
 	"os"
 
-	"github.com/ghodss/yaml"
 	"github.com/linkerd/linkerd2/pkg/k8s"
+	"github.com/linkerd/linkerd2/pkg/k8s/resource"
 	"github.com/spf13/cobra"
 	admissionRegistration "k8s.io/api/admissionregistration/v1beta1"
 	core "k8s.io/api/core/v1"
@@ -15,33 +14,14 @@ import (
 	rbac "k8s.io/api/rbac/v1"
 	apiextension "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	apiRegistration "k8s.io/kube-aggregator/pkg/apis/apiregistration/v1"
 	apiregistrationv1client "k8s.io/kube-aggregator/pkg/client/clientset_generated/clientset/typed/apiregistration/v1"
-
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 )
 
 const (
 	yamlSep = "---\n"
 )
-
-type kubernetesResource struct {
-	runtime.TypeMeta
-	metav1.ObjectMeta `json:"metadata"`
-}
-
-func newKubernetesResource(apiVersion, kind, name string) kubernetesResource {
-	return kubernetesResource{
-		runtime.TypeMeta{
-			APIVersion: apiVersion,
-			Kind:       kind,
-		},
-		metav1.ObjectMeta{
-			Name: name,
-		},
-	}
-}
 
 func newCmdUninstall() *cobra.Command {
 	cmd := &cobra.Command{
@@ -72,34 +52,19 @@ func uninstallRunE(ctx context.Context) error {
 	}
 
 	for _, r := range resources {
-		if err := r.renderResource(os.Stdout); err != nil {
+		if err := r.RenderResource(os.Stdout); err != nil {
 			return fmt.Errorf("error rendering Kubernetes resource:%v", err)
 		}
 	}
 	return nil
 }
 
-func (r kubernetesResource) renderResource(w io.Writer) error {
-	b, err := yaml.Marshal(r)
-	if err != nil {
-		return err
-	}
-
-	_, err = w.Write(b)
-	if err != nil {
-		return err
-	}
-
-	_, err = w.Write([]byte(yamlSep))
-	return err
-}
-
-func fetchKubernetesResources(ctx context.Context, k *k8s.KubernetesAPI) ([]kubernetesResource, error) {
+func fetchKubernetesResources(ctx context.Context, k *k8s.KubernetesAPI) ([]resource.Kubernetes, error) {
 	options := metav1.ListOptions{
 		LabelSelector: k8s.ControllerNSLabel,
 	}
 
-	resources := make([]kubernetesResource, 0)
+	resources := make([]resource.Kubernetes, 0)
 
 	clusterRoles, err := fetchClusterRoles(ctx, k, options)
 	if err != nil {
@@ -161,29 +126,29 @@ func fetchKubernetesResources(ctx context.Context, k *k8s.KubernetesAPI) ([]kube
 	return resources, nil
 }
 
-func fetchClusterRoles(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]kubernetesResource, error) {
+func fetchClusterRoles(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]resource.Kubernetes, error) {
 	list, err := k.RbacV1().ClusterRoles().List(ctx, options)
 	if err != nil {
 		return nil, err
 	}
 
-	resources := make([]kubernetesResource, len(list.Items))
+	resources := make([]resource.Kubernetes, len(list.Items))
 	for i, item := range list.Items {
-		resources[i] = newKubernetesResource(rbac.SchemeGroupVersion.String(), "ClusterRole", item.Name)
+		resources[i] = resource.New(rbac.SchemeGroupVersion.String(), "ClusterRole", item.Name)
 	}
 
 	return resources, nil
 }
 
-func fetchClusterRoleBindings(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]kubernetesResource, error) {
+func fetchClusterRoleBindings(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]resource.Kubernetes, error) {
 	list, err := k.RbacV1().ClusterRoleBindings().List(ctx, options)
 	if err != nil {
 		return nil, err
 	}
 
-	resources := make([]kubernetesResource, len(list.Items))
+	resources := make([]resource.Kubernetes, len(list.Items))
 	for i, item := range list.Items {
-		resources[i] = newKubernetesResource(rbac.SchemeGroupVersion.String(), "ClusterRoleBinding", item.Name)
+		resources[i] = resource.New(rbac.SchemeGroupVersion.String(), "ClusterRoleBinding", item.Name)
 	}
 
 	return resources, nil
@@ -192,89 +157,89 @@ func fetchClusterRoleBindings(ctx context.Context, k *k8s.KubernetesAPI, options
 // Although role bindings are namespaced resources in nature
 // some admin role bindings are created and persisted in the kube-system namespace and will not be deleted
 // when the namespace is deleted
-func fetchKubeSystemRoleBindings(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]kubernetesResource, error) {
+func fetchKubeSystemRoleBindings(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]resource.Kubernetes, error) {
 	list, err := k.RbacV1().RoleBindings("kube-system").List(ctx, options)
 	if err != nil {
 		return nil, err
 	}
 
-	resources := make([]kubernetesResource, len(list.Items))
+	resources := make([]resource.Kubernetes, len(list.Items))
 	for i, item := range list.Items {
-		r := newKubernetesResource(rbac.SchemeGroupVersion.String(), "RoleBinding", item.Name)
+		r := resource.New(rbac.SchemeGroupVersion.String(), "RoleBinding", item.Name)
 		r.Namespace = item.Namespace
 		resources[i] = r
 	}
 	return resources, nil
 }
 
-func fetchCustomResourceDefinitions(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]kubernetesResource, error) {
+func fetchCustomResourceDefinitions(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]resource.Kubernetes, error) {
 	list, err := k.Apiextensions.ApiextensionsV1beta1().CustomResourceDefinitions().List(ctx, options)
 	if err != nil {
 		return nil, err
 	}
 
-	resources := make([]kubernetesResource, len(list.Items))
+	resources := make([]resource.Kubernetes, len(list.Items))
 	for i, item := range list.Items {
-		resources[i] = newKubernetesResource(apiextension.SchemeGroupVersion.String(), "CustomResourceDefinition", item.Name)
+		resources[i] = resource.New(apiextension.SchemeGroupVersion.String(), "CustomResourceDefinition", item.Name)
 	}
 
 	return resources, nil
 }
 
-func fetchNamespaceResource(ctx context.Context, k *k8s.KubernetesAPI) (kubernetesResource, error) {
+func fetchNamespaceResource(ctx context.Context, k *k8s.KubernetesAPI) (resource.Kubernetes, error) {
 	obj, err := k.CoreV1().Namespaces().Get(ctx, controlPlaneNamespace, metav1.GetOptions{})
 	if err != nil {
 		if kerrors.IsNotFound(err) {
-			return kubernetesResource{}, nil
+			return resource.Kubernetes{}, nil
 		}
-		return kubernetesResource{}, err
+		return resource.Kubernetes{}, err
 	}
 
-	return newKubernetesResource(core.SchemeGroupVersion.String(), "Namespace", obj.Name), nil
+	return resource.New(core.SchemeGroupVersion.String(), "Namespace", obj.Name), nil
 }
 
-func fetchPodSecurityPolicy(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]kubernetesResource, error) {
+func fetchPodSecurityPolicy(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]resource.Kubernetes, error) {
 	list, err := k.PolicyV1beta1().PodSecurityPolicies().List(ctx, options)
 	if err != nil {
 		return nil, err
 	}
 
-	resources := make([]kubernetesResource, len(list.Items))
+	resources := make([]resource.Kubernetes, len(list.Items))
 	for i, item := range list.Items {
-		resources[i] = newKubernetesResource(policy.SchemeGroupVersion.String(), "PodSecurityPolicy", item.Name)
+		resources[i] = resource.New(policy.SchemeGroupVersion.String(), "PodSecurityPolicy", item.Name)
 	}
 
 	return resources, nil
 }
 
-func fetchValidatingWebhooksConfiguration(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]kubernetesResource, error) {
+func fetchValidatingWebhooksConfiguration(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]resource.Kubernetes, error) {
 	list, err := k.AdmissionregistrationV1beta1().ValidatingWebhookConfigurations().List(ctx, options)
 	if err != nil {
 		return nil, err
 	}
 
-	resources := make([]kubernetesResource, len(list.Items))
+	resources := make([]resource.Kubernetes, len(list.Items))
 	for i, item := range list.Items {
-		resources[i] = newKubernetesResource(admissionRegistration.SchemeGroupVersion.String(), "ValidatingWebhookConfiguration", item.Name)
+		resources[i] = resource.New(admissionRegistration.SchemeGroupVersion.String(), "ValidatingWebhookConfiguration", item.Name)
 	}
 
 	return resources, nil
 }
 
-func fetchMutatingWebhooksConfiguration(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]kubernetesResource, error) {
+func fetchMutatingWebhooksConfiguration(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]resource.Kubernetes, error) {
 	list, err := k.AdmissionregistrationV1beta1().MutatingWebhookConfigurations().List(ctx, options)
 	if err != nil {
 		return nil, err
 	}
 
-	resources := make([]kubernetesResource, len(list.Items))
+	resources := make([]resource.Kubernetes, len(list.Items))
 	for i, item := range list.Items {
-		resources[i] = newKubernetesResource(admissionRegistration.SchemeGroupVersion.String(), "MutatingWebhookConfiguration", item.Name)
+		resources[i] = resource.New(admissionRegistration.SchemeGroupVersion.String(), "MutatingWebhookConfiguration", item.Name)
 	}
 
 	return resources, nil
 }
-func fetchAPIRegistrationResources(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]kubernetesResource, error) {
+func fetchAPIRegistrationResources(ctx context.Context, k *k8s.KubernetesAPI, options metav1.ListOptions) ([]resource.Kubernetes, error) {
 	apiClient, err := apiregistrationv1client.NewForConfig(k.Config)
 	if err != nil {
 		return nil, err
@@ -285,9 +250,9 @@ func fetchAPIRegistrationResources(ctx context.Context, k *k8s.KubernetesAPI, op
 		return nil, err
 	}
 
-	resources := make([]kubernetesResource, len(list.Items))
+	resources := make([]resource.Kubernetes, len(list.Items))
 	for i, item := range list.Items {
-		resources[i] = newKubernetesResource(apiRegistration.SchemeGroupVersion.String(), "APIService", item.Name)
+		resources[i] = resource.New(apiRegistration.SchemeGroupVersion.String(), "APIService", item.Name)
 	}
 
 	return resources, nil
