+++ multicluster/service-mirror/cluster_watcher_test_util.go
