+++ multicluster/static/templates.go
@@ -6,8 +6,10 @@ package static
 import (
 	"net/http"
 	"path"
+
+	"github.com/linkerd/linkerd2/pkg/charts/static"
 )
 
 // Templates that will be rendered by `linkerd install`. This is only used on
 // dev builds.
-var Templates http.FileSystem = http.Dir(path.Join(GetRepoRoot(), "charts"))
+var Templates http.FileSystem = http.Dir(path.Join(static.GetRepoRoot(), "multicluster/charts"))
